requirements.txt
cocotb
cocotbext-axi==0.1.24
pytest==7.4.3
